NotaryDZ/requirements.txt
PyQt6
sqlalchemy
cryptography
python-dotenv
openai
pyinstaller
